requirements.txt
spacy
transformers
torch
streamlit
pdfplumber
python-docx
scikit-learn
numpy
pandas
sentencepiece
